widrsx-backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
scapy==2.5.0
pyyaml==6.0.1
psutil==5.9.5
